search filter edits
