+++ react-app/src/components/BrowsePets/BrowsePets.css
@@ -1,3 +1,17 @@
+.browse__container {
+    display: flex;
+    align-items: center;
+}
+
+.browse__bar > input {
+    width: 600px;
+    height: 50px;
+}
+
+.search__text {
+    font-size: 14px;
+}
+
 .result__container {
     display: flex;
     flex-wrap: wrap;
