--- react-app/src/components/BrowsePetsBar/BrowsePetsBar.css
@@ -1,13 +1,0 @@
-.browse__container {
-    display: flex;
-    align-items: center;
-}
-
-.browse__bar > input {
-    width: 600px;
-    height: 50px;
-}
-
-.search__text {
-    font-size: 14px;
-}
